requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.0.1
numpy==1.24.3
python-dateutil==2.8.2